% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacex.R
\name{spacex_latest}
\alias{spacex_latest}
\title{Fetch the latest SpaceX launch}
\usage{
spacex_latest()
}
\value{
tibble: single row with same columns as spacex_launches
}
\description{
Fetch the latest SpaceX launch
}
