% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacex.R
\name{spacex_rockets}
\alias{spacex_rockets}
\title{Fetch SpaceX rockets}
\usage{
spacex_rockets()
}
\value{
tibble: id, name, type, active, stages, boosters,
cost_per_launch, first_flight, country, company, description
}
\description{
Fetch SpaceX rockets
}
